+++ nobodywho/Cargo.nix
@@ -9800,6 +9800,10 @@ rec {
             packageId = "futures";
           }
           {
+            name = "log";
+            packageId = "log";
+          }
+          {
             name = "nobodywho";
             packageId = "nobodywho";
           }
@@ -9818,10 +9822,7 @@ rec {
           {
             name = "tracing";
             packageId = "tracing";
-          }
-          {
-            name = "tracing-subscriber";
-            packageId = "tracing-subscriber";
+            features = [ "log" ];
           }
         ];
         buildDependencies = [